refactor: modularize infrastructure tests
